Add graceful shutdown
