+++ docker-context/entrypoint.sh
@@ -1,4 +1,4 @@
 #!/bin/sh
 
 ansible-playbook /root/startup.yml -vv
-server
+exec server
